pyproject.toml
[build-system]
requires = ["flit_core >=3.11,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "scribe"
authors = [{name = "Jay", email = "me@bsavage.xyz"}]
license = "MIT"
license-files = ["LICENSE"]
dynamic = ["version", "description"]

dependencies = [
    "fastapi",
    "uvicorn",
    "aio-pika",
    "pymongo",
    "motor",
    "numpy",
    "pyaudio",
    "faster-whisper",
    "soundfile",
    "pyannote.audio",
    "torchaudio"
]

[tool.flit.module]
name = "scribe"

[project.urls]
Home = "https://github.com/realbucksavage/scribe"

[project.scripts]
run-server = "scribe:run_server"
run-agent = "scribe_agent:run_agent"